+++ src/db/models/token_detail.rs
@@ -6,7 +6,7 @@ pub struct DatabaseTokenDetails {
     pub token: String,
     pub name: String,
     pub symbol: String,
-    pub decimals: i16,           // Only for ERC20 tokens
+    pub decimals: i64,           // Only for ERC20 tokens
     pub token0: Option<String>,  // Only for ERC20 from UniswapV2 Pairs
     pub token1: Option<String>,  // Only for ERC20 from UniswapV2 Pairs
     pub factory: Option<String>, // Only for ERC20 from UniswapV2 Pairs
